Add a fake dependabot-api server
Attempted to set up a fake dependabot-api server using json-server.

The integration test now boots up a localhost server running json-server
that looks similar to dependabot-api allowing the updater to complete
the update.

I haven't fully grocked how it works but should be possible to get it to
persist the pr create from the updater and assert it from the test.