Introduce `integration` test helper to DRY up integration tests
Previously, we'd manually return from blocks that were only ran in tests
marked as `integration` tests, denoted by files that end in
`*-integration.test.ts`.

This lead to a bunch of duplication, as each
block was still started, and we had to mark several blocks in those
files with a piece of code to return early in the case of integration
tests.

This introduces an `integration` helper, which conditionally skips the
entire suite based on wether or not the `INTEGRATION` env var is set.