Use the ImageService to fetch docker dependencies in CI, dev
As a preamble to pinning the image versions we use this introduces
`npm run fetch-images` as a way to pre-pull the images defined in docker_tags.ts
which we will set to specific SHAs in future versions.

This ensures CI and developers pull the images before attempting to run the code
to avoid any surprise breakages.

It also makes the presence of a GITHUB_TOKEN envvar a validation check in
ImageService.pull to avoid confusing docker errors if it isn't present.

Finally, it avoids passing any auth credentials to non-GitHub hosts when
we run our tests

Co-Authored by: Philip Harrison <philip@mailharrison.com>